+++ models/src/anemoi/models/layers/mapper.py
@@ -194,6 +194,7 @@ class GraphTransformerBaseMapper(GraphEdgeMixin, BaseMapper):
         num_chunks: int = 1,
         cpu_offload: bool = False,
         activation: str = "GELU",
+        qk_norm: bool = False,
         num_heads: int = 16,
         mlp_hidden_ratio: int = 4,
         sub_graph: Optional[HeteroData] = None,
@@ -220,6 +221,8 @@ class GraphTransformerBaseMapper(GraphEdgeMixin, BaseMapper):
             ratio of mlp hidden dimension to embedding dimension, default 4
         activation : str, optional
             Activation function, by default "GELU"
+        qk_norm : bool, optional
+            Whether to use query and key normalization, default False
         cpu_offload : bool, optional
             Whether to offload processing to CPU, by default False
         out_channels_dst : Optional[int], optional
@@ -252,6 +255,7 @@ class GraphTransformerBaseMapper(GraphEdgeMixin, BaseMapper):
             num_heads=num_heads,
             edge_dim=self.edge_dim,
             activation=activation,
+            qk_norm=qk_norm,
             num_chunks=num_chunks,
             layer_kernels=layer_kernels,
         )
@@ -308,6 +312,7 @@ class GraphTransformerForwardMapper(ForwardMapperPreProcessMixin, GraphTransform
         num_chunks: int = 1,
         cpu_offload: bool = False,
         activation: str = "GELU",
+        qk_norm: bool = False,
         num_heads: int = 16,
         mlp_hidden_ratio: int = 4,
         sub_graph: Optional[HeteroData] = None,
@@ -334,6 +339,8 @@ class GraphTransformerForwardMapper(ForwardMapperPreProcessMixin, GraphTransform
             ratio of mlp hidden dimension to embedding dimension, default 4
         activation : str, optional
             Activation function, by default "GELU"
+        qk_norm : bool, optional
+            Whether to use query and key normalization, default False
         cpu_offload : bool, optional
             Whether to offload processing to CPU, by default False
         out_channels_dst : Optional[int], optional
@@ -348,6 +355,7 @@ class GraphTransformerForwardMapper(ForwardMapperPreProcessMixin, GraphTransform
             num_chunks=num_chunks,
             cpu_offload=cpu_offload,
             activation=activation,
+            qk_norm=qk_norm,
             num_heads=num_heads,
             mlp_hidden_ratio=mlp_hidden_ratio,
             sub_graph=sub_graph,
@@ -388,8 +396,10 @@ class GraphTransformerBackwardMapper(BackwardMapperPostProcessMixin, GraphTransf
         num_chunks: int = 1,
         cpu_offload: bool = False,
         activation: str = "GELU",
+        qk_norm: bool = False,
         num_heads: int = 16,
         mlp_hidden_ratio: int = 4,
+        initialise_data_extractor_zero: bool = False,
         sub_graph: Optional[HeteroData] = None,
         sub_graph_edge_attributes: Optional[list[str]] = None,
         src_grid_size: int = 0,
@@ -412,8 +422,12 @@ class GraphTransformerBackwardMapper(BackwardMapperPostProcessMixin, GraphTransf
             Number of heads to use, default 16
         mlp_hidden_ratio: int
             ratio of mlp hidden dimension to embedding dimension, default 4
+        initialise_data_extractor_zero : bool, default False:
+            Whether to initialise the data extractor to zero
         activation : str, optional
             Activation function, by default "GELU"
+        qk_norm : bool, optional
+            Whether to use query and key normalization, default False
         cpu_offload : bool, optional
             Whether to offload processing to CPU, by default False
         out_channels_dst : Optional[int], optional
@@ -431,6 +445,7 @@ class GraphTransformerBackwardMapper(BackwardMapperPostProcessMixin, GraphTransf
             num_chunks=num_chunks,
             cpu_offload=cpu_offload,
             activation=activation,
+            qk_norm=qk_norm,
             num_heads=num_heads,
             mlp_hidden_ratio=mlp_hidden_ratio,
             sub_graph=sub_graph,
@@ -443,6 +458,12 @@ class GraphTransformerBackwardMapper(BackwardMapperPostProcessMixin, GraphTransf
         self.node_data_extractor = nn.Sequential(
             nn.LayerNorm(self.hidden_dim), nn.Linear(self.hidden_dim, self.out_channels_dst)
         )
+        if initialise_data_extractor_zero:
+            for module in self.node_data_extractor.modules():
+                if isinstance(module, nn.Linear):
+                    nn.init.constant_(module.weight, 0.0)
+                    if module.bias is not None:
+                        nn.init.constant_(module.bias, 0.0)
 
     def pre_process(self, x, shard_shapes, model_comm_group=None, x_src_is_sharded=False, x_dst_is_sharded=False):
         x_src, x_dst, shapes_src, shapes_dst = super().pre_process(