+++ training/src/anemoi/training/schemas/training.py
@@ -12,6 +12,7 @@ from __future__ import annotations
 from enum import Enum
 from functools import partial
 from typing import Annotated
+from typing import Any
 from typing import Literal
 from typing import Union
 
@@ -20,11 +21,11 @@ from pydantic import Field
 from pydantic import NonNegativeFloat
 from pydantic import NonNegativeInt
 from pydantic import PositiveInt
+from pydantic import field_validator
 from pydantic import model_validator
 
-from anemoi.training.schemas.utils import allowed_values
-
-from .utils import BaseModel
+from anemoi.utils.schemas import BaseModel
+from anemoi.utils.schemas.errors import allowed_values
 
 
 class GradientClip(BaseModel):
@@ -74,8 +75,41 @@ class LR(BaseModel):
     "Number of iterations."
     min: NonNegativeFloat = Field(example=3e-7)
     "Minimum learning rate."
-    warmup_t: NonNegativeInt = Field(example=1000)
+    warmup: NonNegativeInt = Field(example=1000)
     "Number of warm up iteration. Default to 1000."
+
+
+class OptimizerSchema(BaseModel):
+    """Optimizer configuration."""
+
+    zero: bool = Field(example=False)
+    "Use Zero optimiser."
+    kwargs: dict[str, Any] = Field(default_factory=dict)
+    "Additional arguments to pass to the optimizer."
+
+
+class ExplicitTimes(BaseModel):
+    """Time indices for input and output.
+
+    Starts at index 0. Input and output can overlap.
+    """
+
+    input: list[NonNegativeInt] = Field(examples=[0, 1])
+    "Input time indices."
+    target: list[NonNegativeInt] = Field(examples=[2])
+    "Target time indices."
+
+
+class TargetForcing(BaseModel):
+    """Forcing parameters for target output times.
+
+    Extra forcing parameters to use as input to distinguish between different target times.
+    """
+
+    data: list[str] = Field(examples=["insolation"])
+    "List of forcing parameters to use as input to the model at the interpolated step."
+    time_fraction: bool = Field(example=True)
+    "Use target time as a fraction between input boundary times as input."
 
 
 class LossScalingSchema(BaseModel):
@@ -106,7 +140,10 @@ class PressureLevelScalerSchema(BaseModel):
     "Slope of the scaling function."
 
 
-PossibleScalars = Annotated[str, AfterValidator(partial(allowed_values, values=["variable", "loss_weights_mask"]))]
+PossibleScalars = Annotated[
+    str,
+    AfterValidator(partial(allowed_values, values=["limited_area_mask", "variable", "loss_weights_mask", "*"])),
+]
 
 
 class ImplementedLossesUsingBaseLossSchema(str, Enum):
@@ -114,6 +151,10 @@ class ImplementedLossesUsingBaseLossSchema(str, Enum):
     mse = "anemoi.training.losses.mse.WeightedMSELoss"
     mae = "anemoi.training.losses.mae.WeightedMAELoss"
     logcosh = "anemoi.training.losses.logcosh.WeightedLogCoshLoss"
+    huber = "anemoi.training.losses.huber.WeightedHuberLoss"
+    kcrps = "anemoi.training.losses.kcrps.KernelCRPS"
+    afkcrps = "anemoi.training.losses.kcrps.AlmostFairKernelCRPS"
+    limited_mse = "anemoi.training.losses.limitedarea.WeightedMSELossLimitedArea"
 
 
 class BaseLossSchema(BaseModel):
@@ -123,6 +164,19 @@ class BaseLossSchema(BaseModel):
     "Scalars to include in loss calculation"
     ignore_nans: bool = False
     "Allow nans in the loss and apply methods ignoring nans for measuring the loss."
+
+
+class KernelCRPSSchema(BaseLossSchema):
+    fair: bool = True
+    "Calculate a 'fair' (unbiased) score - ensemble variance component weighted by (ens-size-1)^-1"
+
+
+class AlmostFairKernelCRPSSchema(BaseLossSchema):
+    alpha: float = 1.0
+    """Factor for linear combination of fair (unbiased, ensemble variance component
+    weighted by (ens-size-1)^-1) and standard CRPS (1.0 = fully fair, 0.0 = fully unfair)"""
+    no_autocast: bool = True
+    "Deactivate autocast for the kernel CRPS calculation"
 
 
 class HuberLossSchema(BaseLossSchema):
@@ -137,35 +191,96 @@ class WeightedMSELossLimitedAreaSchema(BaseLossSchema):
     "Whether to compute the contribution to the MSE or not."
 
 
-class CombinedLossSchema(BaseModel):
+class CombinedLossSchema(BaseLossSchema):
     target_: Literal["anemoi.training.losses.combined.CombinedLoss"] = Field(..., alias="_target_")
     losses: list[BaseLossSchema] = Field(min_length=1)
-    loss_weights: list[Union[int, float]] = Field(min_length=1)
+    "Losses to combine, can be any of the normal losses."
+    loss_weights: Union[list[Union[int, float]], None] = None
+    "Weightings of losses, if not set, all losses are weighted equally."
+
+    @field_validator("losses", mode="before")
+    @classmethod
+    def add_empty_scalars(cls, losses: Any) -> Any:
+        """Add empty scalars to loss functions, as scalars can be set at top level."""
+        from omegaconf.omegaconf import open_dict
+
+        for loss in losses:
+            if "scalars" not in loss:
+                with open_dict(loss):
+                    loss["scalars"] = []
+        return losses
 
     @model_validator(mode="after")
-    def check_length_of_weights_and_losses(self, values: dict) -> CombinedLossSchema:
-        losses, loss_weights = values["losses"], values["loss_weights"]
-        if len(losses) != len(loss_weights):
+    def check_length_of_weights_and_losses(self) -> CombinedLossSchema:
+        """Check that the number of losses and weights match, or if not set, skip."""
+        losses, loss_weights = self.losses, self.loss_weights
+        if loss_weights is not None and len(losses) != len(loss_weights):
             error_msg = "Number of losses and weights must match"
             raise ValueError(error_msg)
-        return values
+        return self
 
 
-LossSchemas = Union[BaseLossSchema, HuberLossSchema, WeightedMSELossLimitedAreaSchema, CombinedLossSchema]
+LossSchemas = Union[
+    BaseLossSchema,
+    HuberLossSchema,
+    WeightedMSELossLimitedAreaSchema,
+    CombinedLossSchema,
+    KernelCRPSSchema,
+    AlmostFairKernelCRPSSchema,
+]
 
 
-class NodeLossWeightsTargets(str, Enum):
-    graph_node_attribute = "anemoi.training.losses.nodeweights.GraphNodeAttribute"
-    reweighted_graph_node_attributes = "anemoi.training.losses.ReweightedGraphNodeAttribute"
+class ImplementedStrategiesUsingBaseDDPStrategySchema(str, Enum):
+    ddp_ens = "anemoi.training.distributed.strategy.DDPEnsGroupStrategy"
+    ddp = "anemoi.training.distributed.strategy.DDPGroupStrategy"
 
 
-class NodeLossWeightsSchema(BaseModel):
-    target_: NodeLossWeightsTargets = Field(..., alias="_target_")
+class BaseDDPStrategySchema(BaseModel):
+    """Strategy configuration."""
+
+    target_: ImplementedStrategiesUsingBaseDDPStrategySchema = Field(..., alias="_target_")
+    num_gpus_per_model: PositiveInt = Field(example=2)
+    "Number of GPUs per model."
+    read_group_size: PositiveInt = Field(example=1)
+    "Number of GPUs per reader group. Defaults to number of GPUs."
+
+
+class DDPEnsGroupStrategyStrategySchema(BaseDDPStrategySchema):
+    """Strategy object from anemoi.training.strategy."""
+
+    num_gpus_per_ensemble: PositiveInt = Field(example=2)
+    "Number of GPUs per ensemble."
+
+
+StrategySchemas = Union[
+    BaseDDPStrategySchema,
+    DDPEnsGroupStrategyStrategySchema,
+]
+
+
+class GraphNodeAttributeSchema(BaseModel):
+    target_: Literal["anemoi.training.losses.nodeweights.GraphNodeAttribute"] = Field(..., alias="_target_")
     "Node loss weights object from anemoi.training.losses."
     target_nodes: str = Field(examples=["data"])
     "name of target nodes, key in HeteroData graph object."
     node_attribute: str = Field(examples=["area_weight"])
     "name of node weight attribute, key in HeteroData graph object."
+
+
+class ReweightedGraphNodeAttributeSchema(BaseModel):
+    target_: Literal["anemoi.training.losses.nodeweights.ReweightedGraphNodeAttribute"] = Field(..., alias="_target_")
+    "Node loss weights object from anemoi.training.losses."
+    target_nodes: str = Field(examples=["data"])
+    "name of target nodes, key in HeteroData graph object."
+    node_attribute: str = Field(examples=["area_weight"])
+    "name of node weight attribute, key in the nodes object."
+    scaled_attribute: str = Field(examples=["cutout_mask"])
+    "name of node attribute defining the subset of nodes to be scaled, key in the nodes object."
+    weight_frac_of_total: float = Field(examples=[0.3], ge=0, le=1)
+    "sum of weight of subset nodes as a fraction of sum of weight of all nodes after rescaling"
+
+
+NodeLossWeightsSchema = Union[GraphNodeAttributeSchema, ReweightedGraphNodeAttributeSchema]
 
 
 class ScaleValidationMetrics(BaseModel):
@@ -181,7 +296,7 @@ class ScaleValidationMetrics(BaseModel):
     """List of metrics to keep in normalised space.."""
 
 
-class TrainingSchema(BaseModel):
+class BaseTrainingSchema(BaseModel):
     """Training configuration."""
 
     run_id: Union[str, None] = Field(example=None)
@@ -204,14 +319,14 @@ class TrainingSchema(BaseModel):
     accum_grad_batches: PositiveInt = Field(default=1)
     """Accumulates gradients over k batches before stepping the optimizer.
     K >= 1 (if K == 1 then no accumulation). The effective bacthsize becomes num-device * k."""
-    num_sanity_val_steps: PositiveInt = Field(example=6)
+    num_sanity_val_steps: NonNegativeInt = Field(example=6)
     "Sanity check runs n batches of val before starting the training routine."
     gradient_clip: GradientClip
     "Config for gradient clipping."
+    strategy: StrategySchemas
+    "Strategy to use."
     swa: SWA = Field(default_factory=SWA)
     "Config for stochastic weight averaging."
-    zero_optimizer: bool = Field(example=False)
-    "use ZeroRedundancyOptimizer, saves memory for larger models."
     training_loss: LossSchemas
     "Training loss configuration."
     loss_gradient_scaling: bool = False
@@ -228,6 +343,8 @@ class TrainingSchema(BaseModel):
     "Maximum number of steps, stops earlier if max_epochs is reached first."
     lr: LR = Field(default_factory=LR)
     "Learning rate configuration."
+    optimizer: OptimizerSchema = Field(default_factory=OptimizerSchema)
+    "Optimizer configuration."
     variable_loss_scaling: LossScalingSchema
     "Configuration of the variable scaling used in the loss computation."
     pressure_level_scaler: PressureLevelScalerSchema
@@ -236,3 +353,27 @@ class TrainingSchema(BaseModel):
     "List of metrics"
     node_loss_weights: NodeLossWeightsSchema
     "Node loss weights configuration."
+
+
+class ForecasterSchema(BaseTrainingSchema):
+    model_task: Literal["anemoi.training.train.forecaster.GraphForecaster",] = Field(..., alias="model_task")
+    "Training objective."
+
+
+class ForecasterEnsSchema(BaseTrainingSchema):
+    model_task: Literal["anemoi.training.train.forecaster.GraphEnsForecaster",] = Field(..., alias="model_task")
+    "Training objective."
+    ensemble_size_per_device: PositiveInt = Field(example=1)
+    "Number of ensemble member per device"
+
+
+class InterpolationSchema(BaseTrainingSchema):
+    model_task: Literal["anemoi.training.train.forecaster.GraphInterpolator"] = Field(..., alias="model_task")
+    "Training objective."
+    explicit_times: ExplicitTimes
+    "Time indices for input and output."
+    target_forcing: TargetForcing
+    "Forcing parameters for target output times."
+
+
+TrainingSchema = Union[ForecasterSchema, ForecasterEnsSchema, InterpolationSchema]